Began massive changes.  Stage now set for overhaul.
