+++ bin/AndroidManifest.xml
@@ -1,17 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.Centaurii.app.RatingCalculator"
-    android:versionCode="8"
-    android:versionName="1.2.4" >
+    android:versionCode="9"
+    android:versionName="2.0" >
 
     <uses-sdk 
-        android:targetSdkVersion="8"
-        android:minSdkVersion="8">
-     </uses-sdk>
+        android:targetSdkVersion="18"
+        android:minSdkVersion="8"/>
    
+    <supports-screens android:largeScreens="false"/>
     <application
+        android:theme="@style/appTheme"
         android:icon="@drawable/ic_launcher"
-        android:label="@string/appName" >
+        android:label="@string/appName"
+        android:allowBackup="true" >
         <activity
             android:name=".GameRatingCalculatorActivity"
             android:label="@string/appName" 
@@ -23,6 +25,5 @@
             </intent-filter>
         </activity>
     </application>
-    <support-screens android:largeScreens="false"/>
 
 </manifest>
