+++ gen/com/Centaurii/app/RatingCalculator/R.java
@@ -14,82 +14,20 @@ public final class R {
         public static final int ic_launcher=0x7f020000;
     }
     public static final class id {
-        public static final int EditTextp2=0x7f05001e;
-        public static final int EditTextp3=0x7f050022;
-        public static final int EditTextp4=0x7f050026;
-        public static final int EditTextp5=0x7f05002a;
-        public static final int EditTextp6=0x7f05002e;
-        public static final int TableRow01=0x7f050008;
-        public static final int calculateButton=0x7f050031;
-        public static final int checkBox01=0x7f050009;
-        public static final int checkBox02=0x7f05000a;
-        public static final int checkBox03=0x7f05000b;
-        public static final int checkBox04=0x7f05000c;
-        public static final int checkBox05=0x7f05000d;
-        public static final int checkBox06=0x7f05000e;
-        public static final int checkBox1=0x7f050012;
-        public static final int checkBox2=0x7f050013;
-        public static final int checkBox3=0x7f050014;
-        public static final int checkBox4=0x7f050015;
-        public static final int checkBox5=0x7f050016;
-        public static final int checkBox6=0x7f050017;
-        public static final int editTextp1=0x7f05001a;
-        public static final int numPlayersTextView=0x7f050001;
-        public static final int p1Text=0x7f05001b;
-        public static final int p2Text=0x7f05001f;
-        public static final int p3Text=0x7f050023;
-        public static final int p4Text=0x7f050027;
-        public static final int p5Text=0x7f05002b;
-        public static final int p6Text=0x7f05002f;
-        public static final int player1=0x7f050019;
-        public static final int player2=0x7f05001d;
-        public static final int player3=0x7f050021;
-        public static final int player4=0x7f050025;
-        public static final int player5=0x7f050029;
-        public static final int player6=0x7f05002d;
-        public static final int radioGroup2=0x7f050004;
-        public static final int sizeSeekBar=0x7f050005;
-        public static final int tableLayout=0x7f050000;
-        public static final int tableRow1=0x7f050002;
-        public static final int tableRow10=0x7f05002c;
-        public static final int tableRow11=0x7f050030;
-        public static final int tableRow12=0x7f050006;
-        public static final int tableRow2=0x7f050003;
-        public static final int tableRow3=0x7f05000f;
-        public static final int tableRow4=0x7f050011;
-        public static final int tableRow5=0x7f050018;
-        public static final int tableRow6=0x7f05001c;
-        public static final int tableRow7=0x7f050020;
-        public static final int tableRow8=0x7f050024;
-        public static final int tableRow9=0x7f050028;
-        public static final int textView2=0x7f050010;
-        public static final int winnerTextView=0x7f050007;
+        public static final int header=0x7f060001;
+        public static final int main_table=0x7f060000;
     }
     public static final class layout {
-        public static final int main=0x7f030000;
+        public static final int fragment_calculator=0x7f030000;
+        public static final int home_screen=0x7f030001;
+        public static final int main=0x7f030002;
     }
     public static final class string {
-        public static final int Provisional=0x7f04000a;
-        public static final int ProvisionalPlayers=0x7f040009;
-        public static final int appName=0x7f040014;
+        public static final int appName=0x7f040002;
         public static final int app_name=0x7f040000;
-        public static final int calculate=0x7f040013;
-        public static final int five=0x7f040006;
-        public static final int four=0x7f040005;
-        public static final int newRating=0x7f04000d;
-        public static final int numPlayers=0x7f040004;
-        public static final int one=0x7f040008;
-        public static final int player1=0x7f04000c;
-        public static final int player2=0x7f04000e;
-        public static final int player3=0x7f04000f;
-        public static final int player4=0x7f040010;
-        public static final int player5=0x7f040011;
-        public static final int player6=0x7f040012;
-        public static final int six=0x7f040007;
-        public static final int three=0x7f040002;
-        public static final int threePlayer=0x7f040001;
-        public static final int threePlayers=0x7f040003;
-        public static final int two=0x7f04000b;
-        public static final int winningPlayer=0x7f040015;
+        public static final int ratings_calculator=0x7f040001;
+    }
+    public static final class style {
+        public static final int appTheme=0x7f050000;
     }
 }
