+++ project.properties
@@ -11,4 +11,4 @@
 #proguard.config=${sdk.dir}\tools\proguard\proguard-android.txt:proguard-project.txt
 
 # Project target.
-target=android-8
+target=android-18
